GoL-98 fixed casting-bugs

+++ GameOfLife/core/src/main/aau/se2/glock/alpha/gameoflife/screens/BasicScreen.java
@@ -95,7 +95,7 @@ public class BasicScreen implements Screen {
         buttonWidth = screenWidth / 5;
         buttonHeight = screenHeight / 8;
 
-        buttonPosition = new Vector2((float) centerWidth - (buttonWidth / 2), (float) centerHeight - buttonHeight);
+        buttonPosition = new Vector2((float) centerWidth - ((float) buttonWidth / 2), (float) centerHeight - buttonHeight);
     }
 
     /**
